backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
sqlalchemy==2.0.23
pydantic==2.5.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
aiohttp==3.9.1
python-multipart==0.0.6
feedparser==6.0.10